split up update and throttle update on sensor
