+++ homeassistant/components/soma/cover.py
@@ -67,3 +67,23 @@ class SomaCover(SomaEntity, CoverEntity):
     def is_closed(self):
         """Return if the cover is closed."""
         return self.current_position == 0
+
+    async def async_update(self):
+        """Update the cover with the latest data."""
+        try:
+            _LOGGER.debug("Soma Cover Update")
+            response = await self.hass.async_add_executor_job(
+                self.api.get_shade_state, self.device["mac"]
+            )
+        except RequestException:
+            _LOGGER.error("Connection to SOMA Connect failed")
+            self.is_available = False
+            return
+        if response["result"] != "success":
+            _LOGGER.error(
+                "Unable to reach device %s (%s)", self.device["name"], response["msg"]
+            )
+            self.is_available = False
+            return
+        self.current_position = 100 - response["position"]
+        self.is_available = True
